+++ src/components/shared/MobileMenu/WraperLink.tsx
@@ -30,31 +30,29 @@ export const WraperLink = ({
   return (
     <Box
       sx={{
-        margin: '5px 0px 10px 0px',
+        margin: '5px 0px 10px 10px',
+        borderRadius: '16px',
+        ':hover': {
+          backgroundColor: 'rgba(15, 113, 203, 0.05)',
+        },
       }}
       onClick={onClick}
     >
-      <ListItem>
+      <ListItem
+        sx={{
+          paddingLeft: '10px',
+        }}
+      >
         <Link
           className="titleBox"
           sx={{
             color: 'text.primary',
-            pointerEvents: 'auto',
-            display: 'flex',
-            justifyContent: 'space-between',
-            alignItems: 'center',
-            borderRadius: 3,
-            typography: 'label3',
-            textDecoration: 'none',
-            whiteSpace: 'nowrap',
             ':hover': {
-              backgroundColor: 'selected.hover',
               color: 'primary.dark',
               '.MuiSvgIcon-root': {
                 color: 'primary.dark',
               },
             },
-            cursor: 'pointer',
           }}
         >
           {children}