+++ CHANGELOG.md
@@ -37,6 +37,10 @@
 - Hamburger
   - Full height user menu
   - Click outside to close menu
+- Company Wallet
+  - Card style updates for Active, Inactive and Unknown status
+- App Change Description
+  - fixed display of the description correctly
 
 ### Feature